requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14.0
django-allauth>=0.57.0
django-cors-headers>=4.3.0
djangorestframework-simplejwt>=5.3.0
requests>=2.31.0
cryptography>=41.0.0
gunicorn>=21.2.0
whitenoise>=6.6.0
psycopg2-binary>=2.9.7
dj-database-url>=2.1.0 